added support to storage blog resource
