+++ internal/services/storage/blobs.go
@@ -3,6 +3,8 @@ package storage
 import (
 	"bytes"
 	"context"
+	"encoding/base64"
+	"encoding/hex"
 	"fmt"
 	"io"
 	"os"
@@ -360,4 +362,22 @@ func (sbu BlobUpload) blobPageUploadWorker(ctx context.Context, uploadCtx blobPa
 
 		uploadCtx.wg.Done()
 	}
+}
+
+func convertHexToBase64Encoding(str string) (string, error) {
+	data, err := hex.DecodeString(str)
+	if err != nil {
+		return "", fmt.Errorf("converting %q from Hex to Base64 Encoding: %+v", str, err)
+	}
+
+	return base64.StdEncoding.EncodeToString(data), nil
+}
+
+func convertBase64ToHexEncoding(str string) (string, error) {
+	data, err := base64.StdEncoding.DecodeString(str)
+	if err != nil {
+		return "", fmt.Errorf("converting %q from Base64 to Hex Encoding: %+v", str, err)
+	}
+
+	return hex.EncodeToString(data), nil
 }
